Fix leaky array resize and split out command submission
Command buffers were previously bound tightly to the frame they were created for. This meant that in order to record a frame, you first had to wait for the command buffer associated with it to be in available.

Splitting it out allows one to record a buffer while the previous frame's buffer is still processing,. On submission, we wait for the old command buffer to finish before submitting the new one.